se elimina generar-token.js, se crea controllers/profileControllers.js para ordenar mis archivos
