+++ api/.gitignore
@@ -1,3 +1,3 @@
 node_modules/
 .env
-generar-token.js
+/generar-token.js
